refactor: move char guess into user_input module
