+++ src/user_input/user_input.h
@@ -1,7 +1,9 @@
 #include <stdio.h>
+#include "../data_types/my_string/my_string.h"
+#include "../data_types/boolean.h"
 
 int get_num_guesses_input(void);
 
 int get_word_length_input(void);
 
-int get_char_guess_input(void);
+char get_char_guess_input(MY_STRING previous_guesses);
